Correct insertion sort
Using size_t will cause segment fault because the index variable
might become negative during computation.

+++ insertion_sort/src/insertion_sort.cc
@@ -2,9 +2,9 @@
 
 void InsertionSort(std::vector<int>* input)
 {
-    for (size_t i = 1; i < input->size(); ++i) {
+    for (int i = 1; i < input->size(); ++i) {
         int value = (*input)[i];
-        size_t j = i-1;
+        int j = i-1;
         while (j >= 0 && (*input)[j] > value) {
             (*input)[j+1] = (*input)[j];
             --j;
